Update ToyRobot CLI to accept file input by name
Update the Toy Robot CLI module's `main function to accept an optional
filename as input and then process it instead of loading the application
and then accepting User input one at a time. This will allow a User to
predefine a set of instructions for the Robot within a simple text,
which then can be feed into the CLI via the command line to be
processed.

== References:
- [IO and the file system \- Elixir]
  (https://elixir-lang.org/getting-started/io-and-the-file-system.html)

- [IO – Elixir v1\.6\.2]
  (https://hexdocs.pm/elixir/IO.html)

- [Working With the File System in Elixir]
  (https://code.tutsplus.com/tutorials/working-with-file-system-in-elixir--cms-28869)